Fixed segment fault at kuto::memory at end, no more u32 at for loop(because uint32_t isn't the fastest) and fixed xyz loader.
git-svn-id: https://rpgtukuru-iphone.googlecode.com/svn/trunk@66 07b74652-1305-11df-902e-ef6c94960d2c

+++ font/Font.cpp
@@ -37,7 +37,7 @@ const void* Font::getGothic(const char* input0)
 const void* Font::getMincho(const char* input0)
 {
 	const unsigned char* input = (const unsigned char*)input0;
-	if( *input < (u8)0x80 ) {
+	if( *input < 0x80 ) {
 		return HANKAKU[static_cast<int>(*input)];
 	} else if( (0x80 <  *input) && (*input < 0xa0) ) {
 		int ku = (*input - 0x80) * 2, ten;
@@ -71,7 +71,7 @@ bool Font::isFullChar(const char* input0)
 bool Font::isHalfChar(const char* input0)
 {
 	const unsigned char* input = (const unsigned char*)input0;
-	if( ( *input < (u8)0x80 ) ||
+	if( ( *input < 0x80 ) ||
 			( (0xa0 <  *input) && (*input < 0xe0) )
 	) {
 		return true;
